pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "diff-match-patch==20241021",
    "django==5.1.7",
    "django-import-export==4.3.7",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.5.0",
    "drf-spectacular~=0.28.0",
    "drf-yasg>=1.21.10",
    "pandas>=2.2.3",
    "psycopg2-binary==2.9.10",
    "pyjwt==2.9.0",
    "python-dotenv==1.0.1",
    "reportlab>=4.3.1",
    "sqlparse==0.5.3",
    "tablib==3.8.0",
]
